Extract embedded-terminal region into sections/terminal.js
Finishes what P-07 started. The terminal-tab subsystem is now one directory: sections/terminal.js owns the cross-region state (termState) and the helpers tab-drag reaches back in for (_termSideEl, _termTabsOn, _termActiveTab, _termSendResize, _termSendResizeAll, _termPersistTabs, _termShellLabel, _termSyncActiveSide, _termSetActive, _termRenderTabChip, _termAssetsReady, _termWarnAssets, _termClipboardRead, _termClipboardWrite); sections/tab-drag.js already owns the pointer-event drag + tab create/close/rename + shortcuts + screenshot-paste + restore-from-localStorage half. The circular import is now entirely intra-sections (tab-drag.js ↔ terminal.js), not sections ↔ dashboard-main.js, matching the original P-07 plan.

Three other consumers retargeted: tab-drag.js, git-actions.js, runner-viewers.js all now import termState / _termActiveTab / _termSendResize / _termClipboard* from ./terminal.js instead of ../dashboard-main.js. 212 LOC moved.
